refactor create user code
