+++ js/apps/admin-ui/src/user/form-state.ts
@@ -5,18 +5,15 @@ import {
   keyValueToArray,
 } from "../components/key-value-form/key-value-convert";
 import { beerify, debeerify } from "../util";
+import { ToznyPasswordBrokerFieldsForm } from "./ToznyPasswordBrokerFields";
 
 export type UserFormFields = Omit<
   UIUserRepresentation,
   "attributes" | "userProfileMetadata" | "unmanagedAttributes"
-> & {
+> &
+  ToznyPasswordBrokerFieldsForm &{
   attributes?: KeyValueType[] | Record<string, string | string[]>;
   unmanagedAttributes?: KeyValueType[] | Record<string, string | string[]>;
-  brokerUrl?: string
-  authentication?: {
-    emailRecoveryExpirationMinutes?: number;
-    adminRecoveryExpirationMinutes?: number;
-  };
 };
 
 export interface UIUserRepresentation extends UserRepresentation {
